src/pmdas/netbsd: assorted NetBSD 8.0 changes
- misc compiler warning removal, e.g. printf warnings involving
  sizeof() on 64-bit platforms, unused variables, ...
- #include chaining changed, so to get CPU_STATES defined we need
  to explicitly include <sys/sched.h>
- add mem.freemem as a PMNS alias for mem.util.freemem (needed by
  lots of QA tests)

+++ src/pmdas/netbsd/disk.c
@@ -57,12 +57,12 @@ refresh_disk_metrics(void)
 	/* initialization or something has changed */
 	if (ndisk == -1) {
 	    if (pmDebugOptions.appl0)
-		fprintf(stderr, "Info: refresh_disk_metrics: initial ndisk=%d\n", buflen / sizeof(struct io_sysctl));
+		fprintf(stderr, "Info: refresh_disk_metrics: initial ndisk=%d\n", (int)(buflen / sizeof(struct io_sysctl)));
 	    pmdaCacheOp(indomtab[DISK_INDOM].it_indom, PMDA_CACHE_LOAD);
 	}
 	else {
 	if (pmDebugOptions.appl0)
-	    fprintf(stderr, "Info: refresh_disk_metrics: ndisk changed from %d to %d\n", ndisk, buflen / sizeof(struct io_sysctl));
+	    fprintf(stderr, "Info: refresh_disk_metrics: ndisk changed from %d to %d\n", ndisk, (int)(buflen / sizeof(struct io_sysctl)));
 	    pmdaCacheOp(indomtab[DISK_INDOM].it_indom, PMDA_CACHE_INACTIVE);
 	}
 	ndisk = buflen / sizeof(struct io_sysctl);
